feat(article): 두 아티클 서비스에서 공용으로 사용되는 ArticleAuthor 를 Article.types.ts 로 이동
