+++ apps/react-world/src/apis/article/Article.types.ts
@@ -1,0 +1,6 @@
+export interface ArticleAuthor {
+  username: string;
+  bio: string;
+  image: string;
+  following: boolean;
+}
