+++ apps/react-world/src/apis/article/ArticlePreviewService.types.ts
@@ -1,16 +1,11 @@
+import type { ArticleAuthor } from './Article.types';
+
 export interface ArticlePreviewParams {
   offset: number;
   limit: number;
 }
 
 // TODO: 클라이언트에서 좀 더 가독성 높은 선언이 되도록 프로퍼티 이름 변경하고 매핑 로직 작성할 것
-export interface ArticleAuthor {
-  username: string;
-  bio: string;
-  image: string;
-  following: boolean;
-}
-
 export interface ArticlePreviewData {
   slug: string;
   title: string;
